REFACTOR/FIX: major frontend refactoring
This commit starts from a bug report concerning images not loading when used as foot note.

The source of the bug was discourse lazy-loading library, the image is always hidden before the tooltip is shown and as a result has never been tracked by our lazy-loading library. This commit changes the css of the notes container to be hidden (visbility: hidden) and in absolute position instead of not displayed (display: none).

Moreover this plugin applies the following changes:
- removes jquery usage
- uses popper for positioning
- uses event delegation to have only one listener for the footnot click
- cleanly removes event listeners