UX: Fix height issues with inline footnotes
`visibility: hidden` still takes up space, so the footnotes were
causing the height of the post to be larger than it should be in some
cases (tall images in the footnote, or lots of text in it).

This switches to hiding the list and separator. (When footnotes are
inlined, the `footnote-backref` element is inside the list, so it's
already hidden).

See t/98004 internally.